+++ src/Symfony/Component/Messenger/Envelope.php
@@ -26,9 +26,10 @@ final class Envelope
     private $message;
 
     /**
-     * @param object $message
+     * @param object           $message
+     * @param StampInterface[] $stamps
      */
-    public function __construct($message, StampInterface ...$stamps)
+    public function __construct($message, array $stamps = [])
     {
         if (!\is_object($message)) {
             throw new \TypeError(sprintf('Invalid argument provided to "%s()": expected object but got %s.', __METHOD__, \gettype($message)));
@@ -38,6 +39,19 @@ final class Envelope
         foreach ($stamps as $stamp) {
             $this->stamps[\get_class($stamp)][] = $stamp;
         }
+    }
+
+    /**
+     * Makes sure the message is in an Envelope and adds the given stamps.
+     *
+     * @param object|Envelope  $message
+     * @param StampInterface[] $stamps
+     */
+    public static function wrap($message, array $stamps = []): self
+    {
+        $envelope = $message instanceof self ? $message : new self($message);
+
+        return $envelope->with(...$stamps);
     }
 
     /**