+++ src/Symfony/Component/Messenger/Tests/HandleTraitTest.php
@@ -31,7 +31,7 @@ class HandleTraitTest extends TestCase
 
         $query = new DummyMessage('Hello');
         $bus->expects($this->once())->method('dispatch')->willReturn(
-            new Envelope($query, new HandledStamp('result', 'DummyHandler::__invoke'))
+            new Envelope($query, [new HandledStamp('result', 'DummyHandler::__invoke')])
         );
 
         $this->assertSame('result', $queryBus->query($query));
@@ -42,7 +42,7 @@ class HandleTraitTest extends TestCase
         $bus = $this->createMock(MessageBus::class);
         $queryBus = new TestQueryBus($bus);
 
-        $envelope = new Envelope(new DummyMessage('Hello'), new HandledStamp('result', 'DummyHandler::__invoke'));
+        $envelope = new Envelope(new DummyMessage('Hello'), [new HandledStamp('result', 'DummyHandler::__invoke')]);
         $bus->expects($this->once())->method('dispatch')->willReturn($envelope);
 
         $this->assertSame('result', $queryBus->query($envelope));
@@ -74,7 +74,7 @@ class HandleTraitTest extends TestCase
 
         $query = new DummyMessage('Hello');
         $bus->expects($this->once())->method('dispatch')->willReturn(
-            new Envelope($query, new HandledStamp('first_result', 'FirstDummyHandler::__invoke'), new HandledStamp('second_result', 'SecondDummyHandler::__invoke', 'dummy_2'))
+            new Envelope($query, [new HandledStamp('first_result', 'FirstDummyHandler::__invoke'), new HandledStamp('second_result', 'SecondDummyHandler::__invoke', 'dummy_2')])
         );
 
         $queryBus->query($query);