+++ src/Symfony/Component/Messenger/Tests/RoutableMessageBusTest.php
@@ -18,12 +18,13 @@ use Symfony\Component\Messenger\Exception\InvalidArgumentException;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\RoutableMessageBus;
 use Symfony\Component\Messenger\Stamp\BusNameStamp;
+use Symfony\Component\Messenger\Stamp\DelayStamp;
 
 class RoutableMessageBusTest extends TestCase
 {
     public function testItRoutesToTheCorrectBus()
     {
-        $envelope = new Envelope(new \stdClass(), new BusNameStamp('foo_bus'));
+        $envelope = new Envelope(new \stdClass(), [new BusNameStamp('foo_bus')]);
 
         $bus1 = $this->createMock(MessageBusInterface::class);
         $bus2 = $this->createMock(MessageBusInterface::class);
@@ -32,11 +33,12 @@ class RoutableMessageBusTest extends TestCase
         $container->expects($this->once())->method('has')->with('foo_bus')->willReturn(true);
         $container->expects($this->once())->method('get')->will($this->returnValue($bus2));
 
+        $stamp = new DelayStamp(5);
         $bus1->expects($this->never())->method('dispatch');
-        $bus2->expects($this->once())->method('dispatch')->with($envelope)->willReturn($envelope);
+        $bus2->expects($this->once())->method('dispatch')->with($envelope, [$stamp])->willReturn($envelope);
 
         $routableBus = new RoutableMessageBus($container);
-        $this->assertSame($envelope, $routableBus->dispatch($envelope));
+        $this->assertSame($envelope, $routableBus->dispatch($envelope, [$stamp]));
     }
 
     public function testItExceptionOnMissingStamp()
@@ -58,7 +60,7 @@ class RoutableMessageBusTest extends TestCase
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Invalid bus name');
 
-        $envelope = new Envelope(new \stdClass(), new BusNameStamp('foo_bus'));
+        $envelope = new Envelope(new \stdClass(), [new BusNameStamp('foo_bus')]);
 
         $container = $this->createMock(ContainerInterface::class);
         $container->expects($this->once())->method('has')->willReturn(false);